requirements.txt
aenum==3.1.15
aiohttp==3.9.4
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
Flask==2.2.2
flask-cors==6.0.1
frozenlist==1.4.1
future==1.0.0
gunicorn==21.2.0
idna==3.7
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
line-bot-sdk==3.10.2
lxml==5.2.1
MarkupSafe==2.1.5
multidict==6.0.5
packaging==24.0
pydantic==2.7.0
pydantic_core==2.18.1
python-dateutil==2.9.0.post0
requests==2.31.0
six==1.16.0
soupsieve==2.5
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==2.2.2
wrapt==1.16.0
yarl==1.9.4

# Database dependencies
psycopg2-binary==2.9.9
SQLAlchemy==2.0.30
python-dotenv==1.0.1
